python-ml-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
letterboxdpy==1.0.0
scipy==1.11.1
joblib==1.3.2
APScheduler==3.10.4
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.7
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.11.2
matplotlib==3.7.2
seaborn==0.12.2